+++ src/pharmpy/modeling/__init__.py
@@ -2,6 +2,7 @@ from pharmpy.modeling.block_rvs import create_joint_distribution, split_joint_di
 from pharmpy.modeling.common import (
     bump_model_number,
     convert_model,
+    create_config_template,
     generate_model_code,
     get_config_path,
     get_model_covariates,
@@ -237,6 +238,7 @@ __all__ = [
     'cleanup_model',
     'convert_model',
     'create_basic_pk_model',
+    'create_config_template',
     'create_joint_distribution',
     'create_report',
     'create_rng',