+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ New features
 
 * Add function ``modeling.is_linearized``
 * Add function ``modeling.plot_transformed_eta_distributions``
+* Add function ``modeling.create_config_template``
 
 0.91.0 (2023-03-03)
 -------------------